+++ base/applications/mspaint/canvas.h
@@ -52,7 +52,7 @@ public:
     VOID GetImageRect(RECT& rc);
     VOID MoveSelection(INT xDelta, INT yDelta);
     VOID getNewZoomRect(CRect& rcView, INT newZoom, CPoint ptTarget);
-    VOID zoomTo(INT newZoom, INT left = 0, INT top = 0);
+    VOID zoomTo(INT newZoom, LONG left = 0, LONG top = 0);
 
 protected:
     HITTEST m_hitSelection;
